test collasing bis

+++ scripts/gitlab/build_and_test.sh
@@ -12,8 +12,14 @@ project_dir="$(pwd)"
 build_dir="${BUILD_ROOT}/build_${SYS_TYPE}_${COMPILER}"
 hostconfig="${BUILD_ROOT}/uberenv_libs/${COMPILER}.cmake"
 
+# Dependencies
+if [[ "${1}" != "--build-only" && "${1}" != "--test-only" ]]
+then
+    python scripts/uberenv/uberenv.py --spec=${SPEC}
+fi
+
 # Build
-if [[ "${1}" != "--test-only" ]]
+if [[ "${1}" != "--deps-only" && "${1}" != "--test-only" ]]
 then
     rm -rf ${build_dir}
     mkdir -p ${build_dir}
@@ -28,7 +34,7 @@ then
 fi
 
 # Test
-if [[ "${1}" != "--build-only" ]]
+if [[ "${1}" != "--deps-only" && "${1}" != "--build-only" ]]
 then
     if [[ ! -d ${build_dir} ]]
     then
